Criando venda e adicionado itens no carrinho - 2022.10.13.1

+++ lib/ui/widgets/product_card.dart
@@ -7,7 +7,7 @@ import 'package:tech_shop/ui/pages/pages.dart';
 import 'package:tech_shop/ui/temas/temas.dart';
 
 class ProductCard extends StatefulWidget {
-  int preco;
+  double preco;
   int? precoDesconto;
   String nome;
   String descricao;
